+++ modules/video_models/video_run.py
@@ -28,6 +28,8 @@ def generate(*args, **kwargs):
 
     p = processing.StableDiffusionProcessingVideo(
         sd_model=shared.sd_model,
+        video_engine=engine,
+        video_model=model,
         prompt=prompt,
         negative_prompt=negative,
         styles=styles,
@@ -138,6 +140,7 @@ def generate(*args, **kwargs):
     # video_file = images.save_video(p, filename=None, images=processed.images, video_type=video_type, duration=video_duration, loop=video_loop, pad=video_pad, interpolate=video_interpolate) # legacy video save from list of images
     pixels = video_save.images_to_tensor(processed.images)
     _num_frames, video_file = video_save.save_video(
+        p=p,
         pixels=pixels,
         mp4_fps=mp4_fps,
         mp4_codec=mp4_codec,