+++ CHANGELOG.md
@@ -21,6 +21,7 @@
   - fix `wan-2.2-a14b` stage selection  
   - fix `wan-2.2-5b` vae decode  
   - disabling live preview should not disable progress updates  
+  - video tab create `params.txt` with metadata  
 
 ## Update for 2025-10-18